Move osu!catch movement difficulty calculation to an evaluator
Makes the code more in line with the other game modes.
